Changes to support improved media consumption and streamlined topology

+++ scripts/Install-SP2016PreReqs.ps1
@@ -3,7 +3,12 @@
 
     Start-Transcript -Path c:\cfn\log\Install-SP2016PreReqs.ps1.txt -Append
 
-    Start-Process D:\prerequisiteinstaller.exe -ArgumentList '/unattended' -Wait
+    $driveLetter = Get-Volume | ?{$_.DriveType -eq 'CD-ROM'} | select -ExpandProperty DriveLetter
+    if ($driveLetter.Count -gt 1) {
+        throw "More than 1 mounted ISO found"
+    }
+
+    Start-Process "$($driveLetter):\prerequisiteinstaller.exe" -ArgumentList '/unattended' -Wait
 
     Restart-Computer
 }
